add shared utilities for dialogs
