+++ bdot10k_dialog_by_layer.py
@@ -33,15 +33,16 @@ from qgis.core import (Qgis, QgsMessageLog, QgsApplication,
                        QgsCoordinateReferenceSystem)
 from qgis.utils import iface
 
-from .utils import *
+from .dialog_mixin import *
 from .task_dwnl_bdot import DownloadBdotTask
+from .utils import *
 
 # This loads your .ui file so that PyQt can populate your plugin with the elements from Qt Designer
 FORM_CLASS, _ = uic.loadUiType(os.path.join(
     os.path.dirname(__file__), 'bdot10k_dialog_by_layer.ui'))
 
 
-class BDOT10kDialogByLayer(QDialog, FORM_CLASS):
+class BDOT10kDialogByLayer(QDialog, FORM_CLASS, DialogMixin):
     def __init__(self, parent=None):
         """Constructor."""
         super().__init__(parent)
@@ -61,14 +62,6 @@ class BDOT10kDialogByLayer(QDialog, FORM_CLASS):
                                             QgsMapLayerProxyModel.LineLayer |
                                             QgsMapLayerProxyModel.PolygonLayer)
         self.txt.clear()
-
-    def on_gbOldSchema_toggled(self):
-        if self.gbOldSchema.isChecked():
-            self.rbtnGML.setDisabled(True)
-            self.rbtnGPKG.setDisabled(True)
-        else:
-            self.rbtnGML.setDisabled(False)
-            self.rbtnGPKG.setDisabled(False)
 
     def on_mcbLayer_layerChanged(self):
         layerForSelection = self.mcbLayer.currentLayer()
@@ -128,18 +121,7 @@ class BDOT10kDialogByLayer(QDialog, FORM_CLASS):
         else:
             downloadPath = self.dwnlPath.filePath()
 
-            if self.gbOldSchema.isChecked():
-                oldSchema = True
-                if self.rbtnSHPold.isChecked():
-                    bdot10kDataFormat = 'SHP'
-                elif self.rbtnGMLold.isChecked():
-                    bdot10kDataFormat = 'GML'
-            else:
-                oldSchema = False
-                if self.rbtnGML.isChecked():
-                    bdot10kDataFormat = 'GML'
-                elif self.rbtnGPKG.isChecked():
-                    bdot10kDataFormat = 'GPKG'
+            oldSchema, bdot10kDataFormat = self.get_bdot10k_data_options()
 
             if check_dwnl_path(downloadPath):
                 QgsMessageLog.logMessage(f'Lokalizacja pobierania: {downloadPath}', 'BDOT10k', level=Qgis.MessageLevel.Info)
